Cleanup of smartcontracts package on top of ETH-120 PR (#23)
* fix: change compiler version to 0.8.6, remove different openzeppelin versions from package.json

* fix: change all code to use ^0.8.0 compiler, change all licences to MIT

* fix: explicit dependency

although it comes from ENSCache.sol already, too. Explicit is better than implicit.

* chore: move extra test files under a separate dir

* chore: delete unnecessary file

it's not copied under test dir instead of compiled separately. It's only for testing convenience, no need to burden the  build process with it.

* chore: remove chainlink npm packages only used for tests

* fix: remove comments also whitespace fix

* fix: remove zeppelin4 also whitespace fix

* chore: move test/contracts -> test-contracts, because mocha wouldn't otherwise run

problem is there's some ts in there and mocha throws up when it tries to process them for some reason