fix: earningsBeforeJoin -> unpaidEarningsWei
In Data Unions, earningsBeforeJoin are tracked (and called that) because earnings aren't necessarily paid out when leaving. It makes sense to leave tokens into the contract when transfers are expensive (mainnet). Here however the transfer happens always when leaving, so it's not so much "before join" (because that's zero) as it is "unpaid" in the sense that it's recorded in onStakeIncrease but not paid out.

Maybe we should pay out more often? Why not? I guess it might affect some policies e.g. penalizing, so better not go there.