More tests and fixes (#279)
* test: add 8 more tests for insolvency

* refactor: cache localData and globalData

probably saves gas at least if optimizer can't optimize them away? A bit less noise too.

* fix: insolvency end

to see if returning from insolvency, should check if anything was forfeited or not

* fix: debugging oops

* fix: earningsBeforeJoin -> unpaidEarningsWei

In Data Unions, earningsBeforeJoin are tracked (and called that) because earnings aren't necessarily paid out when leaving. It makes sense to leave tokens into the contract when transfers are expensive (mainnet). Here however the transfer happens always when leaving, so it's not so much "before join" (because that's zero) as it is "unpaid" in the sense that it's recorded in onStakeIncrease but not paid out.

Maybe we should pay out more often? Why not? I guess it might affect some policies e.g. penalizing, so better not go there.

* fix: onSponsor update

it has to be done, added comments why

* fix: remove outdated TODO

* test: console.logging improvements

* test: comment out broken assert