Minimum stake (#492)
* refactor: rename agreement -> bounty

the word "agreement" was in early spec and could've referred to "Stream Agreements" (Bounties) and "Broker Agreements" (BrokerPools)

* wip

* test: deployBountyContract means now deploy without BountyFactory

* feat: ETH-450 don't let reduce stake below minimum

* doc: removed outdated comment

* feat: PR review changes: unstaking now throws if would get leave penalty, only "safe" allowed

Also instead of putting the 10/9 into minimumStakeOf, just add it to FLAG_STAKE_WEI and you have the safety margin just fine. Added lots of comments about that.

also the WEIRDEST time-skipping bug in defaultLeavePolicy.test.ts: setting time to +300 actually sets it to +301 for some reason, so I had to set it to +299 to hit +300

* test: speed up VoteKickPolicy tests by deploying only BrokerPoolFactory again

Also added `rewardsBeneficiary` to BrokerPool. Couldn't remove `broker` because it's used for checking what broker's own share in pool is.

* test: fix determinism (again!)

yes, not only factory but also template must be deployed by the clean-slate account. And yes, it's ~4.6% chance one out of 7 doesn't get picked with 20 tries...

* build: fix lint

* build: remove lint warnings

* doc: REVIEWER_SELECTION_ITERATIONS

* test: added test for "not enough stake for flagging"

* test: fix only

* test: add test for "prevent reducing stake too much, so reviewers get paid"

* fix: should be able to flag when stake == minimumStake

* test: rename deployBountyContract -> deployBountyWithoutFactory

* test: fix lint

* feat: minimum stake for all bounties to ensure the kicked broker always has enough stake to slash from

a flagger could lose its stake (almost) arbitrarily low be flagging a lot and paying for flag reviews. Now however the tokens at stake are always calculated from AT LEAST bounty minimumstake, so a sub-minimumstaker would just lose proportionally more than the 10% (because it's 10% of minimum stake now).

Flag stake must have enough "lax" in it so that even after paying for the flags, there's more than 10% of minimum stake to slash.

* fix: slow test

* fix: should be flagger's stake...

* feat: add setter for minimum stake constant

* fix: slow test