Sponsorship fix: don't let _reduceStakeBy revert when over-slashing (#632)
* fix: don't let _reduceStakeBy revert when over-slashing

instead just reduce stake to zero if trying to reduce too much

This case is extremely rare and hard to actually cause; basically if policies commit more stake than the operator has, then _removeOperator might try to _slash more than there is stake. In that case, since it's really "over-slashing" just slash all. It's the responsibility of the policies etc. to make sure there's enough tokens to cover reviewer rewards etc. whatever stake is committed towards.

On the other hand, when capping the slashing to stake, the amount of ACTUAL slashing/reduction must be communicated back, so that the caller doesn't simply assume slashing equals what was requested.

* fix: copy-paste error