ETH-636: Take fisherman's reward from self-delegation (#679)
* feat: fisherman's reward is now paid as % of withdrawn earnings

self-delegation is slashed by burning the corresponding amount of operator tokens. So in net, the operator still makes (fee% - reward%) of the earnings. Which I guess is fair. Like with slashing in sponsorships, if the operator has a fee lower than the `fishermanRewardFraction` and managed to get slashed a lot, down to zero, then the fisherman's reward will be on the delegators.

Also other cleanups:
* reverted to msg.sender where appropriate (still left onlyNodes to use _msgSender even though that 100xs the price of heartbeats probably...)
* reverted some stuff from stake module to main contract. StakeModule size went down by a little at the cost of slightly larger Operator contract (note that StakeModule contains the Operator!)
* added comments, moved some of them back to main contract (just in case someone generates docs from the Operator.sol source...)

* test: improve coverage

also move misplaced tests, deleted duplicate

* refactor: one less argument in a stakemodule interface function

* refactor: more info into custom errors

* test: improve coverage

back to 100% except for that _msgData()...