feat: revert fisherman to service the queue
reason is: we can't have _msgSender() in the module if we want meta-tx to work. So if we want to support fishing with meta-tx (and why not?) then fisherman must get paid in the main contract. So best to just revert it wholesale: it did make the contracts smaller, but also messed up the meta-tx.