OperatorValueUpdate fixes (#696)
* fix: OperatorValueUpdate should happen after withdraw

...and only once after removeSponsorship; nowadays _splitEarnings doesn't call it.

Fisherman also doesn't need to service the queue (it used to be like that before a recent change)

* feat: revert fisherman to service the queue

reason is: we can't have _msgSender() in the module if we want meta-tx to work. So if we want to support fishing with meta-tx (and why not?) then fisherman must get paid in the main contract. So best to just revert it wholesale: it did make the contracts smaller, but also messed up the meta-tx.