Operator-value weighted voting in VoteKickPolicy (#700)
* feat: stake-weighted voting

TODO: subgraph changes

* feat: also save who voted what to the subgraph

* doc: comment Sponsorship entity fields

* fix: naming change

* fix: FlagUpdate handling

sign of voterWeight event param is the vote (kick / no kick)

* fix: handle zero-voter FlagUpdate when flag ends

* feat: add to subgraph Operator.votesOnFlags

* fix: ETH-663 check if stream exists

the check was only on one code path of two, move to common code path

* fix: ignore zero slashing

could slash flagger with zero slashing if no one voted

* test: add test data to dev docker for a completed flag

* fix: move lastFlagIndex from Stake to Flag

now it's kept in the first flag. Stake would be a natural place but Stake entities disappear when the Operator unstakes, so it's not a good place for permanent flag counts

* test: fixed VoteKickPolicy stake locking tests with real StreamrConfig values

* test: fixed all VoteKickPolicy tests with correct StreamrConfig numbers

* test: fixed rest of unit tests too

the problem was MinimumDelegation. Most tests still use "wrong" StreamrConfig values, but they can be fixed piecemeal later.

* feat: updated addresses

new contracts in streamrEnvDeployer caused address changes

* feat: updated subgraph YML addresses

* test: skip DU test

changed addresses makes DU addresses out of date, and updating them would mean deploying a DU subgraph deployer image, probably, and that's in another repo completely. It can't be updated so that it's correct for this PR and master at the same time.

This test probably belongs elsewhere. Or DU deployment would have to be done using different addresses. Probably the whole envdeployer should be refactored a bit. Later.

* fix: smoke test values

* bugfix: remove double deployment of ENS from fastchain image, update ENS-sync-script docker image with new address

* restore old chainlink adapter to original

* feat: move DU deployment before tokenomics

we're expecting the DU to not change much, so at least its addresses won't then change so often

* config fastchain address update

* update ens sync script to changed addresses

* update network-subgraph contract addresses

---------

Co-authored-by: Christoph-Samuel Pitter <christoph-samuel.pitter@streamr.network>