feat: minimum stake calculation
now it's as good as it's ever been: it takes into account if the operator has been flagged or not, so it only reserves the "flag-target stake" if the flagging hasn't been done yet.

How is this different from just locking flag-target stake for everyone?
* targetStakeAtRiskWei depends on stakedWei at the time of flagging, can't know in advance
* locked stake is forfeited when leaving. If flag hasn't been opened, that stake shouldn't be lost, so it shouldn't be locked either (though of course we could at leaving check if there's a flag, and unlock stake if not. BUT now the main contract doesn't need to know anything about flags!)

it could be ever so slightly more elegant if locking also would go into kickpolicy, but then it'd be a bit awkward at the leaving time (removeoperator) vs reducestake where reducestake needs to take into account future flag while removeoperator should not.