feat: removed DEFAULT_ADMIN_ROLE from factories, improved admin role documentation
DEFAULT_ADMIN_ROLE in upgradeable contracts is problematic: it's the default role admin for all roles, so if it leaks to wrong hands, then upgrade won't help: basically the whole OZ RBAC needs to be abandoned.

If instead ADMIN_ROLE leaks (but we still manage to upgrade), we can change that role's ID and the onlyRole modifiers, and thus wipe the role holders list, but otherwise keep the contract intact.