refactor(npm-network-contracts): Capitalize ABI exports (#980)
Export ABIs as capitalized names, e.g. `StreamRegistryABI` instead of
`streamRegistryABI`.

Maybe that naming style is more readable as the ABI is a constant, not a
variable. Could have renamed e.g. to `STREAM_REGISTRY_ABI`, but maybe
the capitalized version makes most sense as it refers to object instead
of an ID.