feat(npm-network-contracts): [ETH-859] DATAv2 contract available via `@streamr/network-contracts` package (#982)
## Background

The `DATAv2` contract implementation is in a separate repo
https://github.com/streamr-dev/DATAv2

## Motivation

For end-user convenience it is good that all our contracts are available
a single package (`@streamr/network-contracts`)

## Implementation

We added `DataV2` contract as a dependency for Hardhat
`dependencyCompiler`. It causes Hardhat to fetch the npm-published
version of the contract and produce build artifacts (ABI and type
definitions) locally for that contract.
 
As a consequence we just needed to add `DATAv2` type and `DATAv2ABI`
export definitions to `exports.ts`.

## Flattening

Ideally we'd just install the `@streamr/data-v2` package and add it to
`dependencyCompiler`, but there were problems with dependency versions.
`@streamr/data-v2` is compiled with `@openzeppelin/contracts` version
4.0.0, and some dependencies of `network-contracts` use 4.9.6; this
isn't a problem to NPM nor to `solc`, but Hardhat throws HH700 errors
that turned tricky to reason about.

Maybe upgrading the version in the token repository could've solved the
problem, but I didn't want to make changes to the token bytecode for
this ticket. But maybe those changes wouldn't really be a problem; made
ticket ETH-862 to manage the token's dependencies once it's moved to the
network-contracts monorepo.

So instead used a flattened version of `DATAv2.sol` from
`@streamr/data-v2`, to avoid these dependency problems.

## Future improvements

We could move `DATAv2` contract implementation to the
`network-contracts` monorepo (ETH-799)