refactor(network-subgraphs): Simplify subgraph yaml management (#1008)
Removed `subgraph.yaml` templating functionality. It was a draft
implementation and was not in use.

Moved build functionality from the npm script to a separate `build.sh`
file. The file requires one parameter which defined the target
environment (e.g. `dev2`).

Renamed `subgraph.yml` to `subgraph_dev2.yml` to highlight that this
file is for `streamr-docker-dev`.

Moved `streamr-docker-dev` scripts to the `start.sh` (which is run when
`deploy-network-subgraphs-fastchain` docker image starts).

Removed some obsolete scripts from package.json (`create-local`,
`deploy-local`, `redeploy-local`)

## Next steps

In a follow-up PRs we'll improve this by:
- modifying `build.sh` so that it build all the output to a `dist`
directory
- reducing copy-paste of between `subgraph.yaml` files by implementing a
better templating functionality

## Future improvements

- Could rename `subgraph_matic.yaml` to `subgraph_polygon.yaml` so that
it communicates the environment clearly.
- The `test` and `coverage` scripts in package.json access some files
from the root level. Could move that functionality to this subpackage.