chore(network-contracts): [ETH-894] Simplify `selectors.txt` generation (#1025)
Simplified the generation of `selectors.txt`. That file is a list of the
[function
selectors](https://docs.soliditylang.org/en/latest/abi-spec.html#function-selector)
of the contracts methods. It also includes selector IDs for contract
errors and events.

As the selector generation is step in the build process, we can easily
export the selectors from the ABI JSONs generated in an earlier step.
The actual calculation of the selector ID is done by `ethers` library:
```ts
ethers.utils.id(signature).slice(0, 10)
```

## Improvements

- Before this PR the selectors didn't contain all IDs of our contracts.
Instead we managed the "top of inheritance-chain contracts" list
indirectly via the imports of `MockRandomOracle.sol`
- Before this PR the file was generated via Docker by using `solc`
compiler. That is no longer needed, and therefore we can generate the
selectors also if Docker is not running.
- the compiler always user v0.8.13 of the compiler, in practice we use
multiple versions in `hardhat.config.js`
- The selector IDs now contain the `0x` prefix. This makes it easier to
search as the error messages typically include the prefix, too.

## Other changes

- Using `npx` to execute `tsc` in the `build.sh` so that the script can
be run directly from command line (instead of via `npm run build`)