refactor(dev-chain-fast): [ETH-895] Move deployers to `dev-chain-fast` (#1027)
Moved `StreamrEnvDeployer` and `HubEnvDeployer` classes from
`network-contracts` and `hub-contracts` subpackages to the
`dev-chain-fast` subpackage. These classes implements the environment
initialization logic for the development environment.

As the actual development environment initialization is done in in
`dev-chain-fast` subpackage (which is the exact purpose of that
subpackage), it makes sense to move the files there.

## Exports

This changes the exports of `@streamr/hub-contracts` but doesn't change
the exports of `@streamr/network-contracts` because the
`network-contracts` package is published from a different subpackage
(`npm-network-subgraphs`). The actual contracts don't change, just the
`HubEnvDeployer` is removed from the exports. That class may be
considered as an internal class and therefore this is not a breaking
change. We'll collect all contacts to a separate npm package in the
future (ETH-799), and that change will deprecate both of these packages.

## Future improvements

We could merge `HubEnvDeployer` to be part of `StreamEnvDeployer`